+++ RELEASENOTES.md
@@ -27,6 +27,9 @@
     *   Fix bug where silence skipping at the end of items can trigger a
         playback exception.
     *   Add `clear` to `PreloadMediaSource` to discard the preloading period.
+    *   Add new error code
+        `PlaybackException.ERROR_CODE_DECODING_RESOURCES_RECLAIMED` that is used
+        when codec resources are reclaimed for higher priority tasks.
 *   Transformer:
     *   Work around a decoder bug where the number of audio channels was capped
         at stereo when handling PCM input.